+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## Next
+- Add `Target` struct helper for creating targets, and hide the targets in the `target` module.
+
 ## 0.1.0 - 2023 Jan 11
 - Initial release